backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
automata-lib>=7.0.0
graphviz==0.20.1
python-multipart>=0.0.5
openai==1.3.0
regex>=2023.0.0
transformers>=4.30.0
python-dotenv==1.0.0
gunicorn>=21.0.0  # For production deployment
httpx>=0.24.0  # Required for FastAPI testing and async requests
pytest>=7.0.0  # For testing
redis==5.0.1
flask==3.0.0
flask-cors==4.0.0
python-json-logger==2.0.7
requests==2.31.0
aiofiles==23.2.1
lxml>=4.9.3
aioredis>=2.0.1
pydantic>=2.0.0
structlog>=23.1.0
slowapi>=0.1.8  # For rate limiting
prometheus-client>=0.17.0  # For metrics
python-jose[cryptography]>=3.3.0  # For JWT handling
passlib[bcrypt]>=1.7.4  # For password hashing
